前日始値終値を使った動的グリッド幅にOperationを追加し、上書き以外に加算や積算にも対応できるよう変更

+++ grid_service.go
@@ -213,7 +213,7 @@ func (s *gridService) width(strategy *Strategy) (int, error) {
 	if strategy.GridStrategy.DynamicGridPrevDay.Valid {
 		fp, err := s.fourPriceStore.GetLastBySymbolCodeAndExchange(strategy.SymbolCode, strategy.Exchange)
 		if err == nil {
-			w = strategy.GridStrategy.DynamicGridPrevDay.width(w, float64(s.tick.Ticks(strategy.TickGroup, fp.Open, fp.Close)))
+			w = strategy.GridStrategy.DynamicGridPrevDay.width(w, s.tick.Ticks(strategy.TickGroup, fp.Open, fp.Close))
 		}
 	}
